requirements.txt
accelerate==1.8.1
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
attrs==25.3.0
bleach==6.2.0
blinker==1.9.0
blis==1.3.0
braceexpand==0.1.7
cachetools==6.1.0
catalogue==2.0.10
certifi==2025.6.15
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
clip==0.2.0
cloudpathlib==0.21.1
confection==0.1.5
contexttimer==0.3.3
contourpy==1.3.2
cycler==0.12.1
cymem==2.0.11
decorator==5.2.1
decord==0.6.0
distlib==0.3.9
einops==0.8.1
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl#sha256=293e9547a655b25499198ab15a525b05b9407a75f10255e405e8c3854329ab63
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
exceptiongroup==1.3.0
executing==2.2.0
fairscale==0.4.4
faiss-cpu==1.11.0
filelock==3.18.0
fonttools==4.58.4
fsspec==2025.5.1
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
hf-xet==1.1.5
huggingface-hub==0.33.1
identify==2.6.12
idna==3.10
ijson==3.4.0
imageio==2.37.0
iopath==0.1.10
ipython==8.37.0
isodate==0.7.2
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kaggle==1.7.4.5
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.13
narwhals==1.44.0
networkx==3.3
nltk==3.9.1
nodeenv==1.9.1
numpy==2.2.6
nvidia-cublas-cu12==12.8.3.14
nvidia-cuda-cupti-cu12==12.8.57
nvidia-cuda-nvrtc-cu12==12.8.61
nvidia-cuda-runtime-cu12==12.8.57
nvidia-cudnn-cu12==9.7.1.26
nvidia-cufft-cu12==11.3.3.41
nvidia-cufile-cu12==1.13.0.11
nvidia-curand-cu12==10.3.9.55
nvidia-cusolver-cu12==11.7.2.55
nvidia-cusparse-cu12==12.5.7.53
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.8.55
omegaconf==2.3.0
openai-clip==1.0.1
opencv-python==4.11.0.86
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
packaging==25.0
pandas==2.3.0
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.8
plotly==6.1.2
portalocker==3.2.0
pre_commit==4.2.0
preshed==3.0.10
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pycocoevalcap==1.2
pycocotools==2.0.10
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-magic==0.4.27
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
rdflib==7.1.4
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rpds-py==0.25.1
safetensors==0.5.3
salesforce-lavis==1.0.2
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.15.3
sentence-transformers==4.1.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SPARQLWrapper==2.0.0
srsly==2.5.1
stack-data==0.6.3
streamlit==1.46.0
sympy==1.13.3
tenacity==9.1.2
text-unidecode==1.3
thinc==8.3.6
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.15
tokenizers==0.21.2
toml==0.10.2
torch==2.7.1+cu128
torchaudio==2.7.1+cu128
torchvision==0.22.1+cu128
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
triton==3.3.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.31.2
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.13
weasel==0.4.1
webdataset==1.0.2
webencodings==0.5.1
wikitextparser==0.56.4
wrapt==1.17.2
